requirements.txt
discord.py==1.7.3
python-dotenv==0.7.1
pymongo[srv]==4.0.1
